# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "AccountActionHistoryResponse": {
        "additionalProperties": false,
        "description": "Persisted Action history only; non-Action projections never appear here.",
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionDetailResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceResponse"
          }
        },
        "required": [
          "workspace",
          "account_id",
          "items",
          "limit",
          "offset"
        ],
        "title": "AccountActionHistoryResponse",
        "type": "object"
      },
...
      "AccountStateDetailResponse": {
        "additionalProperties": false,
        "description": "One snapshot with normalized explanations and full input traces.",
        "properties": {
          "evaluator_manifest": {
            "$ref": "#/components/schemas/StateEvaluatorManifestResponse"
          },
          "fit_criteria": {
            "items": {
              "$ref": "#/components/schemas/FitCriterionTraceResponse"
            },
            "title": "Fit Criteria",
            "type": "array"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/AccountStateReasonResponse"
            },
            "title": "Reasons",
            "type": "array"
          },
          "relationship_evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceResponse"
            },
            "title": "Relationship Evidence",
            "type": "array"
          },
          "signal_evaluations": {
            "items": {
              "$ref": "#/components/schemas/SignalEvaluationTraceResponse"
            },
            "title": "Signal Evaluations",
            "type": "array"
          },
          "snapshot": {
            "$ref": "#/components/schemas/AccountStateSnapshotResponse"
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceResponse"
          }
        },
        "required": [
          "workspace",
          "snapshot",
          "evaluator_manifest",
          "reasons",
          "fit_criteria",
          "relationship_evidence",
          "signal_evaluations"
        ],
        "title": "AccountStateDetailResponse",
        "type": "object"
      },
...
      "ActionDetailResponse": {
        "additionalProperties": false,
        "description": "Action proposal with projected governance and operational history.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionResponse"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/ActionAttemptTraceResponse"
            },
            "title": "Attempts",
            "type": "array"
          },
          "external_execution_authorized": {
            "const": false,
            "default": false,
            "title": "External Execution Authorized",
            "type": "boolean"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/ActionLifecycle"
          },
          "mutations_enabled": {
            "title": "Mutations Enabled",
            "type": "boolean"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionReviewResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action",
          "lifecycle",
          "review",
          "attempts",
          "mutations_enabled"
        ],
        "title": "ActionDetailResponse",
        "type": "object"
      },
      "ActionDryRunMutationResponse": {
        "additionalProperties": false,
        "description": "Result of local deterministic validation without external execution.",
        "properties": {
          "external_execution_authorized": {
            "const": false,
            "default": false,
            "title": "External Execution Authorized",
            "type": "boolean"
          },
          "idempotent_replay": {
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "trace": {
            "$ref": "#/components/schemas/ActionAttemptTraceResponse"
          }
        },
        "required": [
          "trace",
          "idempotent_replay"
        ],
        "title": "ActionDryRunMutationResponse",
        "type": "object"
      },
      "ActionLifecycle": {
        "description": "Read-only governance projection; never stored as mutable Action state.",
        "enum": [
          "REVIEW_REQUIRED",
          "READY_FOR_DRY_RUN",
          "REJECTED"
        ],
        "title": "ActionLifecycle",
        "type": "string"
      },
      "ActionOutcomeListResponse": {
        "additionalProperties": false,
        "description": "Immutable operational Outcomes for one Action.",
        "properties": {
          "action_id": {
            "format": "uuid",
            "title": "Action Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionOutcomeResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "action_id",
          "items"
        ],
        "title": "ActionOutcomeListResponse",
        "type": "object"
      },
...
      "ActionReviewMutationResponse": {
        "additionalProperties": false,
        "description": "Result of an explicit local-demo review command.",
        "properties": {
          "external_execution_authorized": {
            "const": false,
            "default": false,
            "title": "External Execution Authorized",
            "type": "boolean"
          },
          "idempotent_replay": {
            "title": "Idempotent Replay",
            "type": "boolean"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/ActionLifecycle"
          },
          "review": {
            "$ref": "#/components/schemas/ActionReviewResponse"
          }
        },
        "required": [
          "review",
          "lifecycle",
          "idempotent_replay"
        ],
        "title": "ActionReviewMutationResponse",
        "type": "object"
      },
...
      "ActionReviewRequest": {
        "additionalProperties": false,
        "description": "Bounded unauthenticated-demo review command.",
        "properties": {
          "reason_code": {
            "$ref": "#/components/schemas/ActionReviewReasonCode"
          },
          "resolution": {
            "$ref": "#/components/schemas/ActionReviewResolution"
          }
        },
        "required": [
          "resolution",
          "reason_code"
        ],
        "title": "ActionReviewRequest",
        "type": "object"
      },
...
      "ActionTraceResponse": {
        "additionalProperties": false,
        "description": "Composed end-to-end trace without stored provenance duplication.",
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActionDetailResponse"
          },
          "state": {
            "$ref": "#/components/schemas/AccountStateDetailResponse"
          },
          "strategy": {
            "$ref": "#/components/schemas/StrategyResponse"
          },
          "upstream": {
            "$ref": "#/components/schemas/DecisionPolicyDetailResponse"
          }
        },
        "required": [
          "strategy",
          "state",
          "upstream",
          "action"
        ],
        "title": "ActionTraceResponse",
        "type": "object"
      },
...
      "CurrentAccountActionResponse": {
        "additionalProperties": false,
        "description": "Current exact Decision/Policy disposition and its M1D projection.",
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "current": {
            "$ref": "#/components/schemas/CurrentActionProjectionResponse"
          },
          "upstream": {
            "$ref": "#/components/schemas/DecisionPolicyDetailResponse"
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceResponse"
          }
        },
        "required": [
          "workspace",
          "account_id",
          "upstream",
          "current"
        ],
        "title": "CurrentAccountActionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Typed service boundary for evidence-backed GTM decisions.",
    "title": "GTM State & Signal Engine API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/accounts/{account_id}/action": {
      "get": {
        "description": "Return the current M1D projection separately from persisted Action history.",
        "operationId": "get_current_account_action_api_v1_accounts__account_id__action_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentAccountActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Account Action",
        "tags": [
          "M1D governed Action and Outcome"
        ]
      }
    },
    "/api/v1/accounts/{account_id}/actions": {
      "get": {
        "description": "Return persisted Actions only; BLOCK and abstention projections are excluded.",
        "operationId": "list_account_action_history_api_v1_accounts__account_id__actions_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountActionHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Account Action History",
        "tags": [
          "M1D governed Action and Outcome"
        ]
      }
    },
...
    "/api/v1/accounts/{account_id}/state": {
      "get": {
        "description": "Return the latest materialized snapshot for the semantic state_as_of.",
        "operationId": "get_current_account_state_api_v1_accounts__account_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountStateDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Account State",
        "tags": [
          "M1B.2 account-state read model"
        ]
      }
    },
...
    "/api/v1/actions/{action_id}": {
      "get": {
        "description": "Return one immutable Action proposal and its projected governance state.",
        "operationId": "get_action_api_v1_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Action",
        "tags": [
          "M1D governed Action and Outcome"
        ]
      }
    },
    "/api/v1/actions/{action_id}/dry-runs": {
      "post": {
        "description": "Run local deterministic validation; never simulate external execution.",
        "operationId": "dry_run_action_api_v1_actions__action_id__dry_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "title": " Body",
                "type": "null"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDryRunMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dry Run Action",
        "tags": [
          "M1D governed Action and Outcome"
        ]
      }
    },
    "/api/v1/actions/{action_id}/outcomes": {
      "get": {
        "description": "Return operational dry-run Outcomes only.",
        "operationId": "list_action_outcomes_api_v1_actions__action_id__outcomes_get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOutcomeListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Action Outcomes",
        "tags": [
          "M1D governed Action and Outcome"
        ]
      }
    },
    "/api/v1/actions/{action_id}/reviews": {
      "post": {
        "description": "Record one bounded local-demo review without claiming authenticated identity.",
        "operationId": "review_action_api_v1_actions__action_id__reviews_post",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionReviewMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Action",
        "tags": [
          "M1D governed Action and Outcome"
        ]
      }
    },
    "/api/v1/actions/{action_id}/trace": {
      "get": {
        "description": "Compose Strategy through Outcome without adding redundant provenance rows.",
        "operationId": "get_action_trace_api_v1_actions__action_id__trace_get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionTraceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Action Trace",
        "tags": [
          "M1D governed Action and Outcome"
        ]
      }
    },
...
  }
}